Convert print statements to verbose().

+++ kiosk_content_retrieval/_content-retrieval.php
@@ -26,54 +26,57 @@ if (empty($_REQUEST)) {
   exit();
 }
 else {
+  if (array_key_exists('group', $_REQUEST)) {
   
-  header('Content-type: text/plain');
+    // Get the group name
+    $group_name = $_REQUEST['group'];
   
-  // Get the group name
-  $group_name = $_REQUEST['group'];
-  
-  // Make sure a folder exists for placeing the results
-  $results_directory = 'results/results_for_'.$group_name; 
-  if (!file_exists($results_directory)) {
-    mkdir($results_directory);
-  }
-  
-  // Process all the packages in the chosen group
-  $group = $packages[$group_name];
-  foreach ($group as $package) {
-    foreach ($package['handlers'] as $handler) {
-
-      // Make sure the folder exists for placing the results
-      // TODO: Abstract checking & making directories into a generic function
-      // TODO: Place this call with the respective handler function
-      $handler_directory = $results_directory.'/packs_'.$handler; 
-      if (!file_exists($handler_directory)) {
-        mkdir($handler_directory);
-      }
-      
-      // Run the handler
-      $function = 'process_'.$handler;
-      $data_source = $package['data_source_url']; 
-      $function($data_source, $results_directory);
-      
-      // Add a directory tree (_files.json)
-      $files_json = folder_array($handler_directory);
-      $files_json = json_encode($files_json);
-      file_put_contents($handler_directory.'/_files.json', $files_json);
-      
+    // Make sure a folder exists for placing the results
+    $results_directory = 'results/results_for_'.$group_name; 
+    if (!file_exists($results_directory)) {
+      mkdir($results_directory);
     }
-  }
-  
-  /*
-  // Zip up the package
-  // COMMENTED OUT, BECAUSE COMPRESSION CAN OCCUR FOR THE ENTIRE GAME...NOT JUST THE PACKS
-  // TODO: MAKE THIS AN OPTION
-  if (PHP_OS == 'WINNT') {
-    `compact $results_directory`;
+    
+    // Process all the packages in the chosen group
+    $group = $packages[$group_name];
+    foreach ($group as $package) {
+      foreach ($package['handlers'] as $handler) {
+
+        // Make sure the folder exists for placing the results
+        // TODO: Abstract checking & making directories into a generic function
+        // TODO: Place this call with the respective handler function
+        $handler_directory = $results_directory.'/packs_'.$handler; 
+        if (!file_exists($handler_directory)) {
+          mkdir($handler_directory);
+        }
+        
+        // Run the handler
+        $function = 'process_'.$handler;
+        $data_source = $package['data_source_url']; 
+        $function($data_source, $results_directory);
+        
+        // Add a directory tree (_files.json)
+        $files_json = folder_array($handler_directory);
+        $files_json = json_encode($files_json);
+        file_put_contents($handler_directory.'/_files.json', $files_json);
+        
+      }
+    }
+    
+    /*
+    // Zip up the package
+    // COMMENTED OUT, BECAUSE COMPRESSION CAN OCCUR FOR THE ENTIRE GAME...NOT JUST THE PACKS
+    // TODO: MAKE THIS AN OPTION
+    if (PHP_OS == 'WINNT') {
+      `compact $results_directory`;
+    }
+    else {
+      `tar -cvzf $results_directory.tar.gz $results_directory`;
+    }
+    */
+    
   }
   else {
-    `tar -cvzf $results_directory.tar.gz $results_directory`;
+    // Do nothing...don't work
   }
-  */
-  
 }
